[ARM] Size reduce teq to eors
Add t2TEQrr to the map of instructions with can be reduced down into
a T1 instruction. This is a special case because TEQ just sets the
CPSR and doesn't write to a GPR, which is not the case for EOR. So,
we need to ensure that the EOR can write to the first operand.

Differential Revision: https://reviews.llvm.org/D56255


git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@350801 91177308-0d34-0410-b5e6-96231b3b80d8