[DAGCombiner] fix crash when converting build vector to shuffle
The regression test is reduced from the example shown in D56281.
This does raise a question as noted in the test file: do we want
to handle this pattern? I don't have a motivating example for
that on x86 yet, but it seems like we could have that pattern 
there too, so we could avoid the back-and-forth using a shuffle.


git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@351753 91177308-0d34-0410-b5e6-96231b3b80d8