[x86] split more 256/512-bit shuffles in lowering
This is intentionally a small step because it's hard to know exactly 
where we might introduce a conflicting transform with the code that 
tries to form wider shuffles. But I think this is safe - if we have 
a wide shuffle with 2 operands, then we should do better with an 
extract + narrow shuffle.

Differential Revision: https://reviews.llvm.org/D57867


git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@353427 91177308-0d34-0410-b5e6-96231b3b80d8