AMDGPU: Ignore CodeObjectV3 when inlining
This was inhibiting inlining of library functions when clang was
invoking the inliner directly. This is covering a bit of a mess with
subtarget feature handling, and this shouldn't be a subtarget
feature. The behavior is different depending on whether you are using
a -mattr flag in clang, or llc, opt.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@353899 91177308-0d34-0410-b5e6-96231b3b80d8