[llvm-cxxfilt] Re-enable the delimiters test on Windows.
The original intent was to enable this test for Windows; however, that
initial patch broke one of the build-bots.  I temporarily disabled this
test on Windows until that issue was resolved.  It was resolved in my
previous patch (cfd1d9742ee2d1b8dd6b7), and now I am re-enabling this
test.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@355011 91177308-0d34-0410-b5e6-96231b3b80d8