[x86] eliminate movddup of horizontal op
This pattern would show up as a regression if we more
aggressively convert vector FP ops to scalar ops.

There's still a missed optimization for the v4f64 legal
case (AVX) because we create that h-op with an undef operand.
We should probably just duplicate the operands for that
pattern to avoid trouble.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@357642 91177308-0d34-0410-b5e6-96231b3b80d8