[AArch64][GlobalISel] Fix a crash when selecting shufflevectors with an undef mask element.
If a shufflevector's mask vector has an element with "undef" then the generic
instruction defining that element register is a G_IMPLICT_DEF instead of G_CONSTANT.
This fixes the selector to handle this case, and for now assumes that undef just means
zero. In future we'll optimize this case properly.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@358312 91177308-0d34-0410-b5e6-96231b3b80d8