gn build: Add check-clangd target after r3359424
r359527 already merged some of that to the GN build,
but it was missing some bits as well.

The check-clangd target works (at least for now) differently than all
the other check-foo targets, see https://reviews.llvm.org/D61187

For that reason, there's no gni file and the generated lit configs are
not (yet?) added to llvm-lit/BUILD.gn.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@359570 91177308-0d34-0410-b5e6-96231b3b80d8