[Support] Fix unit test for fs::is_local
Close the temporary file after the test is done using it.
If it is not closed and the file was created on NFS, it will cause the test
to fail. The problem happens in the cleanup process afterwards. It first
tries to delete the file but it is not really deleted. Afterwards, the
program fails to delete the directory containing the file, causing the whole
test to fail.

Patch by Milos Stojanovic.


git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@360259 91177308-0d34-0410-b5e6-96231b3b80d8