Another attempt to fix the build bot breaks after r360426
The test case checks were produced by the update_test_checks.py
scripts and I assumed that is sufficient. However, the behaviour
is different with different default target triples. Specify the
triple explicitly in the test case.

If this doesn't clean up the build bot breaks, I'll remove the test
case until I can get to the bottom of why the behaviour on build bots
is different from my machine.


git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@360434 91177308-0d34-0410-b5e6-96231b3b80d8